feat: separated entities, valueObjects and useCases into new folders
