+++ src/adapter/driver/api/controllers/cliente/controller.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from "express";
-import { ClienteUseCase } from "@useCases/cliente";
+import { ClienteUseCase } from "useCases/cliente";
 import { StatusCode } from "@utils/statusCode";
-import { ClienteDTO } from "@useCases/cliente/dto";
+import { ClienteDTO } from "useCases/cliente/dto";
 
 export class ClienteController {
     constructor(private readonly clienteUseCase: ClienteUseCase) {}
