+++ src/adapter/driver/api/controllers/produto/factory.ts
@@ -1,6 +1,6 @@
 import { ProdutoModel } from "@infra/mongo/models";
 import { ProdutoMongoRepository } from "@infra/mongo/repositories";
-import { ProdutoUseCase } from "@useCases/produto";
+import { ProdutoUseCase } from "useCases/produto";
 import { ProdutoController } from "./controller";
 
 export class ProdutoControllerFactory {
