+++ src/useCases/pedido/pedido.interface.ts
