+++ src/useCases/produto/produto.interface.ts
