+++ src/valueObjects/valorTotal.ts
